Notes from Subversion r6308-r6350
r6350
Fix red color on RX62N-35QT. Bad Chris.
---------------------
r6349
Fix Crossworks parse issue Update Sourceforge script mailsend copy portion. add ramdisk unmount script to repo
---------------------
r6348
Remove LPC43XX_IAP from being built in Crossworks library to match IAR library. This fixes a warning in Crossworks project maker project.
---------------------
r6347
This time actually fix the HEW compile bugs.
---------------------
r6346
Some compilers care what line things are on so redo the network compile fix.
---------------------
r6345
Fix KEIL and HEW build errors with network.h
---------------------
r6344
Pack enums in network types to prevent them from getting mixed up from 8 bit to 32 bit. This would cause the settings structure to shift 3 bytes when copying or sending pointers. Update 4357-50WVN platform file to add some missing includes ...
---------------------
r6343
Fix LPC1768 IAR library, update crossworks lib build script to match others.
---------------------
r6342
Remove more old references to renamed KEIL and Crossworks folders.
---------------------
r6341
Update watchdog doxygen example.
---------------------
r6340
Finally switch all KEIL projects over to C99. Remove old KEIL4.x projects.
---------------------
r6339
Fix more compiler warnings, attempt to fix batch build missing CLANG in LPC4357 Crossworks library project.
---------------------
r6338
Fix library build scripts for new Crossworks variable. Simplify CmdLineSettings.bat.
---------------------
r6337
1st attempt at fixing 35QT build error. Need to verify correct color setting.
---------------------
r6336
Update uEZ version number to 2.09. Fix multiple compile warnings that Crossworks 4 spit out (some only on CLANG).
---------------------
r6335
Remove unneeded old extra header files from older emWin versions. These are for implementing LCD drivers, segment lcd driver, etc. but we have LCD drivers in uEZ already.
---------------------
r6334
Change email list from Chris and John to Alex and Brian. Enjoy the spam. -Bill
---------------------
r6333
Make LPC1788 library for IAR actually be set to LPC1788 instead of LPC4088. (note that has been this way since uEZ 2.04 release with 4088-43WQN uEZGUI) The previous setting had the side effect of setting the SPIFI register clock, among othe...
---------------------
r6332
Updated video conversion to handle Synergy videos
---------------------
r6331
Remove auto-calibration setting from register setup as we confirmed from NDH that it always runs at boot-up. Update comments with new info from NHD. Verify that 4.3M model is the high bright 4.3 PCAP.
---------------------
r6330
Update FocalTech FT53XX touch driver to support being able to detect firmware versions then change the touch threshold based on that. I have included standard values for 3 different firmware versions. Currently the driver only changes the m...
---------------------
r6329
Merge in updates from Synergy Driver for touch detection Use smaller write buffer. Read less bytes per touch. Change touch detection to differentiate between pen up, pen down, and hold. Before we were ignoring pen down, and reporting holds ...
---------------------
r6328
Final I2C adjustment to meet fast mode spec on uEZ-GW1
---------------------
r6327
Slight I2C adjustment as high period is too long on some LPC43XXX boards.
---------------------
r6326
Changes for rev 2 gw
---------------------
r6325
Increased the number of characters for the Command console to 256.
---------------------
r6324
Removed enum is int.
---------------------
r6323
Added Renesas EEPROM configuration.
---------------------
r6322
Updated get tick count call to be interrupt save for Ubiquios.
---------------------
r6321
Added logic to allow for not setting the Chip Select.
---------------------
r6320
Added I2C EEPROM for MAC address storage.
---------------------
r6319
Turned of Ethernet startup debugging.
---------------------
r6318
Turned off BSOD as it does not work properly anyway.
---------------------
r6317
Added GW1 platform files.
---------------------
r6316
Added missing sck pin to fix compile issue.
---------------------
r6315
New Tianma display driver. Added LPC546xx processor file and IAR specific files. Some drivers are working but not all have been setup.
---------------------
r6314
Added new processor and display defines.
---------------------
r6313
Initial library project commit.
---------------------
r6312
Update to FATFS V0.12B. Tested on LPC1788 with slideshow and draw apps and seems to work fine.
---------------------
r6311
Update to use REM to comment out unused line. Add top level batch scripts as they currently are.
---------------------
r6310
Fix wrong MCU name in some 1788 scripts. Add new JLink V612A. Update 4088/4357 IAR scripts to use new JLink and standardized programming routine for internal and SPIFI flash.
---------------------
r6309
Add J-Link scripts to repo as these are getting really complicated.
---------------------
r6308
Fix host name so checkout scripts work again.
---------------------

+++ uEZ/Source/Library/Network/GainSpan/CmdLib/GainSpan_SPI.c
@@ -215,28 +215,28 @@ static void GainSpan_SPI_ProcessIncomingChar(char c)
         if (c == GAINSPAN_SPI_CHAR_INACTIVE_LINK) {
             G_GainSpan_SPI_IsLinkActive = false;
             storeChar = false;
-        } else if (c == GAINSPAN_SPI_CHAR_INACTIVE_LINK2) {
+        } else if (((uint8_t) c) == GAINSPAN_SPI_CHAR_INACTIVE_LINK2) {
             G_GainSpan_SPI_IsLinkActive = false;
             storeChar = false;
         } else {
             // Got an active link worth of data
             G_GainSpan_SPI_IsLinkActive = true;
-            if (c == GAINSPAN_SPI_CHAR_ESC) {
+            if (((uint8_t) c) == GAINSPAN_SPI_CHAR_ESC) {
                 /* Don't use this character, go into escape mode */
                 G_GainSpan_SPI_EscapeCode = true;
                 storeChar = false;
-            } else if (c == GAINSPAN_SPI_CHAR_IDLE) {
+            } else if (((uint8_t) c) == GAINSPAN_SPI_CHAR_IDLE) {
                 storeChar = false;
-            } else if (c == GAINSPAN_SPI_CHAR_UNKNOWN_F1) {
+            } else if (((uint8_t) c) == GAINSPAN_SPI_CHAR_UNKNOWN_F1) {
                 storeChar = false;
-            } else if (c == GAINSPAN_SPI_CHAR_UNKNOWN_F4) {
+            } else if (((uint8_t) c) == GAINSPAN_SPI_CHAR_UNKNOWN_F4) {
                 storeChar = false;
-            } else if (c == GAINSPAN_SPI_CHAR_UNKNOWN_F7) {
+            } else if (((uint8_t) c) == GAINSPAN_SPI_CHAR_UNKNOWN_F7) {
                 storeChar = false;
-            } else if (c == GAINSPAN_SPI_CHAR_FLOW_CONTROL_ON) {
+            } else if (((uint8_t) c) == GAINSPAN_SPI_CHAR_FLOW_CONTROL_ON) {
                 G_GainSpan_SPI_CanTransmit = true;
                 storeChar = false;
-            } else if (c == GAINSPAN_SPI_CHAR_FLOW_CONTROL_OFF) {
+            } else if (((uint8_t) c) == GAINSPAN_SPI_CHAR_FLOW_CONTROL_OFF) {
                 G_GainSpan_SPI_CanTransmit = false;
                 storeChar = false;
             }
